Replace empty with new_empty in nn.functional.pad
This makes it so that we can now vmap over nn.functional.pad (circular
variant). Previously we could not because we were effectively doing
`out.copy_(input)` where the out was created with empty.

This also has the added side effect of cleaning up the code.

Test Plan:
- I tested this using functorch.vmap and can confirm that vmap now
works.
- Unfortunately this doesn't work with the vmap in core so I cannot add
a test for this here.

ghstack-source-id: 432ea5e5e035fac01ea394bd11ef3ad190d978e6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68565