dbr quant overhead[5/x]: remove unnecessary asserts
Summary:

Removes asserts which are duplicate (the same condition is checked
when calculating the hook type, so there is no need to check it again).
For the assert in `validate_is_at_last_seen_idx`, rewrites it to
raise an Error instead to ensure it does not get stripped in
production environments.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

ghstack-source-id: f0b8b17be6ac4140fe4109021dc58da35d51615a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68370